bertopic/requirements.txt
torch>=2.0.0
pydantic<2.0.0
bertopic
scikit-learn
numpy
plotly
hdbscan>=0.8.29
umap-learn>=0.5.3
pandas>=2.0.0
nltk>=3.8.0
